# openapi.yaml
openapi: 3.0.3
info:
  title: FitSync2 Backend API
  version: 0.1.0
  description: Initial contract generated from current backend controllers and DTOs.
servers:
  - url: http://localhost:8080
    description: Local
security:
  - bearerAuth: []
...
paths:
...
  /api/users:
    post:
      tags: [User]
      security: []
      summary: Create user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        '201':
          description: Created
          headers:
            Location:
              description: URI of created resource
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '409':
          $ref: '#/components/responses/ErrorResponse'
  /api/users/exists/{loginId}:
    get:
      tags: [User]
      security: []
      summary: Check loginId existence
      parameters:
        - name: loginId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Exists or not
          content:
            application/json:
              schema:
                type: boolean
  /api/users/me:
    get:
      tags: [User]
      summary: Get my user id
      responses:
        '200':
          description: My user id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
    delete:
      tags: [User]
      summary: Delete current user (soft delete)
      responses:
        '204':
          description: Deleted
        '403':
          $ref: '#/components/responses/ErrorResponse'
...
  /api/routines:
    post:
      tags: [Routine]
      summary: Create routine
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoutineRequest'
      responses:
        '201':
          description: Created
          headers:
            Location:
              description: URI of created resource
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoutineResponse'
    get:
      tags: [Routine]
      summary: Get routine list
      parameters:
        - name: writerId
          in: query
          required: false
          schema:
            type: integer
            format: int64
        - name: ownerId
          in: query
          required: false
          schema:
            type: integer
            format: int64
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/size'
        - $ref: '#/components/parameters/sort'
      responses:
        '200':
          description: Paged list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageRoutineListResponse'
  /api/routines/{routineId}:
    get:
      tags: [Routine]
      summary: Get routine detail
      parameters:
        - name: routineId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Routine detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoutineDetailResponse'
    patch:
      tags: [Routine]
      summary: Update routine
      parameters:
        - name: routineId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoutineUpdateRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoutineResponse'
    delete:
      tags: [Routine]
      summary: Delete routine
      parameters:
        - name: routineId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoutineResponse'
  /api/workouts:
    post:
      tags: [Workout]
      summary: Create workout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkoutRequest'
      responses:
        '201':
          description: Created
          headers:
            Location:
              description: URI of created resource
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkoutResponse'
    get:
      tags: [Workout]
      summary: Get workout list
      parameters:
        - name: ownerId
          in: query
          required: false
          schema:
            type: integer
            format: int64
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/size'
        - $ref: '#/components/parameters/sort'
      responses:
        '200':
          description: Paged list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageWorkoutListResponse'
  /api/workouts/{id}:
    get:
      tags: [Workout]
      summary: Get workout detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Workout detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkoutDetailResponse'
...
  /api/ai/log/routine/{id}:
    get:
      tags: [AI]
      summary: Get AI log detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: AI log
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AILogResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
    size:
      name: size
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 5
    sort:
      name: sort
      in: query
      required: false
      schema:
        type: string
        example: id,desc
  responses:
    ErrorResponse:
      description: Error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ValidationError'
      required: [code, message]
    ValidationError:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
      required: [field, message]
    UserRequest:
      type: object
      properties:
        loginId:
          type: string
        password:
          type: string
        name:
          type: string
        roleType:
          $ref: '#/components/schemas/UserRoleType'
        email:
          type: string
        isSocial:
          type: boolean
        socialProviderType:
          $ref: '#/components/schemas/SocialProviderType'
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
      required: [id]
...
    RoutineRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
        writerId:
          type: integer
          format: int64
        ownerId:
          type: integer
          format: int64
        displayOrder:
          type: integer
        description:
          type: string
          maxLength: 1000
        routineExercises:
          type: array
          items:
            $ref: '#/components/schemas/RoutineExerciseRequest'
      required: [name, writerId, ownerId, displayOrder, description, routineExercises]
    RoutineUpdateRequest:
      type: object
      properties:
        name:
          type: string
        displayOrder:
          type: integer
        description:
          type: string
        newExercises:
          type: array
          items:
            $ref: '#/components/schemas/RoutineExerciseRequest'
        updateExercises:
          type: array
          items:
            $ref: '#/components/schemas/RoutineExerciseUpdateRequest'
        deleteExercises:
          type: array
          items:
            $ref: '#/components/schemas/RoutineExerciseDeleteRequest'
      required: [name, displayOrder, description]
    RoutineExerciseRequest:
      type: object
      properties:
        exerciseId:
          type: integer
          format: int64
        displayOrder:
          type: integer
        description:
          type: string
          maxLength: 1000
        routineSets:
          type: array
          items:
            $ref: '#/components/schemas/RoutineSetRequest'
      required: [exerciseId, displayOrder, description, routineSets]
    RoutineExerciseUpdateRequest:
      type: object
      properties:
        id:
          type: integer
          format: int64
        displayOrder:
          type: integer
        description:
          type: string
          maxLength: 1000
        newRoutineSets:
          type: array
          items:
            $ref: '#/components/schemas/RoutineSetRequest'
        updateRoutineSets:
          type: array
          items:
            $ref: '#/components/schemas/RoutineSetUpdateRequest'
        deleteRoutineSets:
          type: array
          items:
            $ref: '#/components/schemas/RoutineSetDeleteRequest'
      required: [id, displayOrder, description]
    RoutineExerciseDeleteRequest:
      type: object
      properties:
        id:
          type: integer
          format: int64
      required: [id]
    RoutineSetRequest:
      type: object
      properties:
        displayOrder:
          type: integer
        weightKg:
          type: integer
        reps:
          type: integer
        distanceM:
          type: integer
        durationSec:
          type: integer
        speedKmh:
          type: integer
        rpe:
          type: integer
        restTimeSec:
          type: integer
      required: [displayOrder]
    RoutineSetUpdateRequest:
      type: object
      properties:
        id:
          type: integer
          format: int64
        displayOrder:
          type: integer
        weightKg:
          type: integer
        reps:
          type: integer
        distanceM:
          type: integer
        durationSec:
          type: integer
        speedKmh:
          type: integer
        rpe:
          type: integer
        restTimeSec:
          type: integer
      required: [id, displayOrder]
    RoutineSetDeleteRequest:
      type: object
      properties:
        id:
          type: integer
          format: int64
      required: [id]
    RoutineResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
      required: [id]
    RoutineListResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        writerId:
          type: integer
          format: int64
        ownerId:
          type: integer
          format: int64
        displayOrder:
          type: integer
        routineExercises:
          type: array
          items:
            $ref: '#/components/schemas/RoutineExerciseListResponse'
      required: [id, name, writerId, ownerId, displayOrder, routineExercises]
    RoutineExerciseListResponse:
      type: object
      properties:
        name:
          type: string
        displayOrder:
          type: integer
      required: [name, displayOrder]
    RoutineDetailResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        displayOrder:
          type: integer
        description:
          type: string
        routineExercises:
          type: array
          items:
            $ref: '#/components/schemas/RoutineExerciseDetailResponse'
      required: [id, name, displayOrder, description, routineExercises]
    RoutineExerciseDetailResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        exerciseId:
          type: integer
          format: int64
        exerciseName:
          type: string
        displayOrder:
          type: integer
        description:
          type: string
        routineSets:
          type: array
          items:
            $ref: '#/components/schemas/RoutineSetDetailResponse'
      required: [id, exerciseId, exerciseName, displayOrder, description, routineSets]
    RoutineSetDetailResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        displayOrder:
          type: integer
        weightKg:
          type: integer
        reps:
          type: integer
        distanceM:
          type: integer
        durationSec:
          type: integer
        speedKmh:
          type: integer
        rpe:
          type: integer
        restTimeSec:
          type: integer
      required: [id, displayOrder]
    WorkoutRequest:
      type: object
      properties:
        writerId:
          type: integer
          format: int64
        ownerId:
          type: integer
          format: int64
        memo:
          type: string
        workoutExercises:
          type: array
          items:
            $ref: '#/components/schemas/WorkoutExerciseRequest'
      required: [writerId, ownerId, workoutExercises]
    WorkoutExerciseRequest:
      type: object
      properties:
        exerciseId:
          type: integer
          format: int64
        memo:
          type: string
        sets:
          type: array
          items:
            $ref: '#/components/schemas/WorkoutSetRequest'
      required: [exerciseId, sets]
    WorkoutSetRequest:
      type: object
      properties:
        memo:
          type: string
        displayOrder:
          type: integer
        weightKg:
          type: integer
        reps:
          type: integer
        distanceM:
          type: integer
        durationSec:
          type: integer
        speedKmh:
          type: integer
        rpe:
          type: integer
        restTimeSec:
          type: integer
    WorkoutResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
      required: [id]
    WorkoutListResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
      required: [id, createdAt]
    WorkoutDetailResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        writerId:
          type: integer
          format: int64
        ownerId:
          type: integer
          format: int64
        memo:
          type: string
        createdAt:
          type: string
          format: date-time
        workoutExercises:
          type: array
          items:
            $ref: '#/components/schemas/WorkoutExerciseDetailResponse'
      required: [id, writerId, ownerId, createdAt, workoutExercises]
    WorkoutExerciseDetailResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        exercise:
          $ref: '#/components/schemas/ExerciseSummaryResponse'
        memo:
          type: string
        workoutSets:
          type: array
          items:
            $ref: '#/components/schemas/WorkoutSetDetailResponse'
      required: [id, exercise, workoutSets]
    WorkoutSetDetailResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        memo:
          type: string
        weightKg:
          type: integer
        reps:
          type: integer
        distanceM:
          type: integer
        durationSec:
          type: integer
        speedKmh:
          type: integer
        rpe:
          type: integer
        restTimeSec:
          type: integer
      required: [id]
    ExerciseSummaryResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
      required: [id, name]
...
    AILogResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        requestId:
          type: string
        userId:
          type: integer
          format: int64
        model:
          $ref: '#/components/schemas/AIModel'
        feature:
          type: string
        version:
          type: string
        inputContent:
          type: object
          additionalProperties: true
        outputContent:
          type: object
          additionalProperties: true
        inputTokens:
          type: integer
          format: int64
        outputTokens:
          type: integer
          format: int64
        requestTime:
          type: string
          format: date-time
        responseTime:
          type: string
          format: date-time
        elapsedMs:
          type: integer
          format: int64
        status:
          $ref: '#/components/schemas/CallStatus'
        statusMessage:
          type: string
        userFeedBack:
          $ref: '#/components/schemas/FeedBackStatus'
        userFeedBackReason:
          type: string
        userAction:
          $ref: '#/components/schemas/UserAction'
      required: [id, requestId, model, feature, inputContent, outputContent]
...
    PageRoutineListResponse:
      allOf:
        - $ref: '#/components/schemas/PageBase'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/RoutineListResponse'
    PageWorkoutListResponse:
      allOf:
        - $ref: '#/components/schemas/PageBase'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/WorkoutListResponse'
    PageBase:
      type: object
      properties:
        content:
          type: array
          items: {}
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        size:
          type: integer
        number:
          type: integer
        numberOfElements:
          type: integer
        first:
          type: boolean
        last:
          type: boolean
        empty:
          type: boolean
      required: [content, totalElements, totalPages, size, number, numberOfElements, first, last, empty]
...
    UserRoleType:
      type: string
      enum: [MEMBER, TRAINER, ADMIN]
    SocialProviderType:
      type: string
      enum: [NAVER, GOOGLE, KAKAO]
...
    AIModel:
      type: string
      enum: [GPT_4_1_MINI]
    CallStatus:
      type: string
      enum: [SUCCESS, EXCEPTION, FAILURE]
    FeedBackStatus:
      type: string
      enum: [LIKE, DISLIKE, NONE]
    UserAction:
      type: string
      enum: [SAVE_IMMEDIATE, SAVE_LATER, SKIP_SAVE]